Disable -Wextra-semi-stmt for skqp
Bug: http://b/123023349

https://skia-review.googlesource.com/c/skia/+/173761 resolves the
warnings in upstream skia.  If my understanding that skqp is a branch of
skia is correct, the above CL is the one to be cherry-picked to Android,
but it doesn't apply cleanly to the AOSP sources.  Also, unlike skia,
skqp seems to have frequent merges from upstream.

I don't want to cherry-pick it (and resolve conflicts) to AOSP and
induce further pain during a future merge.  Instead, let's just disable
the warning in Android.bp.

Test: mmma external/skqp with the new compiler.
Change-Id: I086c45e45e2e7268cfdc0e4ceafd86c4a5f37507